Extract return-sinking compiler fixtures
Move the coherent return-sinking sample cluster into a focused file and replace its obsolete end-of-file placement note.
